Extract parallel operations in `rustc_data_structures::sync` into a new `parallel` submodule
